requirements.txt
annotated-types==0.7.0
beautifulsoup4==4.13.4
certifi==2025.7.14
charset-normalizer==3.4.2
cloudscraper==1.2.71
dacite==1.9.2
dnspython==2.8.0
email-validator==2.3.0
idna==3.10
pydantic==2.12.4
pydantic_core==2.41.5
pyparsing==3.2.3
python-dotenv==1.2.1
requests==2.32.4
requests-toolbelt==1.0.0
soupsieve==2.7
typing-inspection==0.4.2
typing_extensions==4.14.1
urllib3==2.5.0
